♻️ refactor: share LiteXML command identities across bundles via a pure module
The package ships two independently-bundled entries — the browser build
(`index`/`react`/`renderer`) and the node build (`headless`). Each previously
inlined its own `createCommand(...)` call for the LiteXML commands, so a command
obtained from one entry and dispatched onto an editor registered by the other
silently no-ops (Lexical matches command listeners by object reference, not by
the string label).

Isolate every LiteXML command identity (+ the `DiffAction` enum) into a new
side-effect-free `plugins/litexml/command/symbols.ts`, re-exported from the
existing command modules so nothing else changes. Emit it as a dedicated chunk
in the node build and expose it through a new `@lobehub/editor/litexml-commands`
subpath, so:

- both bundles import the same emitted `symbols.js` → one runtime identity;
- the commands can be imported on the server without pulling the DOM-dependent
  editor bundle (the module only imports `createCommand` from `lexical`).

This lets consumers dispatch `LITEXML_*_COMMAND` directly on a `HeadlessEditor`
kernel instead of routing through wrapper methods.
